+++ src/braket/aws/aws_device.py
@@ -76,6 +76,7 @@ class AwsDevice(Device):
         "Cz": "CZ",
         "Cphaseshift": "CPhaseShift",
         "Xy": "XY",
+        "Iswap": "ISwap",
     }
 
     def __init__(