+++ src/braket/circuits/observables.py
@@ -30,20 +30,28 @@ from braket.circuits.quantum_operator_helpers import (
     verify_quantum_operator_matrix_dimensions,
 )
 from braket.circuits.serialization import IRType, OpenQASMSerializationProperties
-from braket.registers.qubit_set import QubitSet
+from braket.registers import QubitInput, QubitSet, QubitSetInput
 
 
 class H(StandardObservable):
     """Hadamard operation as an observable."""
 
-    def __init__(self):
-        """Examples:
-        >>> Observable.H()
+    def __init__(self, target: QubitInput | None = None):
+        """Initializes Hadamard observable.
+
+        Args:
+            target (QubitInput | None): The target qubit to measure the observable on.
+                If not provided, this needs to be provided from the enclosing result type.
+                Default: `None`.
+
+        Examples:
+            >>> observables.H(0)
+            >>> observables.H()
         """
-        super().__init__(ascii_symbols=["H"])
+        super().__init__(ascii_symbols=["H"], target=target)
 
     def _unscaled(self) -> StandardObservable:
-        return H()
+        return H(self._targets)
 
     def _to_jaqcd(self) -> list[str]:
         if self.coefficient != 1:
@@ -54,9 +62,10 @@ class H(StandardObservable):
         self, serialization_properties: OpenQASMSerializationProperties, target: QubitSet = None
     ) -> str:
         coef_prefix = f"{self.coefficient} * " if self.coefficient != 1 else ""
-        if target:
-            qubit_target = serialization_properties.format_target(int(target[0]))
-            return f"{coef_prefix}h({qubit_target})"
+        targets = target or self._targets
+        qubit_target = int(targets[0]) if targets else None
+        if qubit_target is not None:
+            return f"{coef_prefix}h({serialization_properties.format_target(qubit_target)})"
         return f"{coef_prefix}h all"
 
     def to_matrix(self) -> np.ndarray:
@@ -75,14 +84,22 @@ Observable.register_observable(H)
 class I(Observable):  # noqa: E742
     """Identity operation as an observable."""
 
-    def __init__(self):
-        """Examples:
-        >>> Observable.I()
+    def __init__(self, target: QubitInput | None = None):
+        """Initializes Identity observable.
+
+        Args:
+            target (QubitInput | None): The target qubit to measure the observable on.
+                If not provided, this needs to be provided from the enclosing result type.
+                Default: `None`.
+
+        Examples:
+            >>> observables.I(0)
+            >>> observables.I()
         """
-        super().__init__(qubit_count=1, ascii_symbols=["I"])
+        super().__init__(qubit_count=1, ascii_symbols=["I"], targets=target)
 
     def _unscaled(self) -> Observable:
-        return I()
+        return I(self._targets)
 
     def _to_jaqcd(self) -> list[str]:
         if self.coefficient != 1:
@@ -93,9 +110,11 @@ class I(Observable):  # noqa: E742
         self, serialization_properties: OpenQASMSerializationProperties, target: QubitSet = None
     ) -> str:
         coef_prefix = f"{self.coefficient} * " if self.coefficient != 1 else ""
-        if target:
-            qubit_target = serialization_properties.format_target(int(target[0]))
-            return f"{coef_prefix}i({qubit_target})"
+
+        targets = target or self._targets
+        qubit_target = int(targets[0]) if targets else None
+        if qubit_target is not None:
+            return f"{coef_prefix}i({serialization_properties.format_target(qubit_target)})"
         return f"{coef_prefix}i all"
 
     def to_matrix(self) -> np.ndarray:
@@ -124,14 +143,22 @@ Observable.register_observable(I)
 class X(StandardObservable):
     """Pauli-X operation as an observable."""
 
-    def __init__(self):
-        """Examples:
-        >>> Observable.X()
+    def __init__(self, target: QubitInput | None = None):
+        """Initializes Pauli-X observable.
+
+        Args:
+            target (QubitInput | None): The target qubit to measure the observable on.
+                If not provided, this needs to be provided from the enclosing result type.
+                Default: `None`.
+
+        Examples:
+            >>> observables.X(0)
+            >>> observables.X()
         """
-        super().__init__(ascii_symbols=["X"])
+        super().__init__(ascii_symbols=["X"], target=target)
 
     def _unscaled(self) -> StandardObservable:
-        return X()
+        return X(self._targets)
 
     def _to_jaqcd(self) -> list[str]:
         if self.coefficient != 1:
@@ -142,9 +169,11 @@ class X(StandardObservable):
         self, serialization_properties: OpenQASMSerializationProperties, target: QubitSet = None
     ) -> str:
         coef_prefix = f"{self.coefficient} * " if self.coefficient != 1 else ""
-        if target:
-            qubit_target = serialization_properties.format_target(int(target[0]))
-            return f"{coef_prefix}x({qubit_target})"
+
+        targets = target or self._targets
+        qubit_target = int(targets[0]) if targets else None
+        if qubit_target is not None:
+            return f"{coef_prefix}x({serialization_properties.format_target(qubit_target)})"
         return f"{coef_prefix}x all"
 
     def to_matrix(self) -> np.ndarray:
@@ -161,14 +190,22 @@ Observable.register_observable(X)
 class Y(StandardObservable):
     """Pauli-Y operation as an observable."""
 
-    def __init__(self):
-        """Examples:
-        >>> Observable.Y()
+    def __init__(self, target: QubitInput | None = None):
+        """Initializes Pauli-Y observable.
+
+        Args:
+            target (QubitInput | None): The target qubit to measure the observable on.
+                If not provided, this needs to be provided from the enclosing result type.
+                Default: `None`.
+
+        Examples:
+            >>> observables.Y(0)
+            >>> observables.Y()
         """
-        super().__init__(ascii_symbols=["Y"])
+        super().__init__(ascii_symbols=["Y"], target=target)
 
     def _unscaled(self) -> StandardObservable:
-        return Y()
+        return Y(self._targets)
 
     def _to_jaqcd(self) -> list[str]:
         if self.coefficient != 1:
@@ -179,9 +216,10 @@ class Y(StandardObservable):
         self, serialization_properties: OpenQASMSerializationProperties, target: QubitSet = None
     ) -> str:
         coef_prefix = f"{self.coefficient} * " if self.coefficient != 1 else ""
-        if target:
-            qubit_target = serialization_properties.format_target(int(target[0]))
-            return f"{coef_prefix}y({qubit_target})"
+        targets = target or self._targets
+        qubit_target = int(targets[0]) if targets else None
+        if qubit_target is not None:
+            return f"{coef_prefix}y({serialization_properties.format_target(qubit_target)})"
         return f"{coef_prefix}y all"
 
     def to_matrix(self) -> np.ndarray:
@@ -198,14 +236,22 @@ Observable.register_observable(Y)
 class Z(StandardObservable):
     """Pauli-Z operation as an observable."""
 
-    def __init__(self):
-        """Examples:
-        >>> Observable.Z()
+    def __init__(self, target: QubitInput | None = None):
+        """Initializes Pauli-Z observable.
+
+        Args:
+            target (QubitInput | None): The target qubit to measure the observable on.
+                If not provided, this needs to be provided from the enclosing result type.
+                Default: `None`.
+
+        Examples:
+            >>> observables.Z(0)
+            >>> observables.Z()
         """
-        super().__init__(ascii_symbols=["Z"])
+        super().__init__(ascii_symbols=["Z"], target=target)
 
     def _unscaled(self) -> StandardObservable:
-        return Z()
+        return Z(self._targets)
 
     def _to_jaqcd(self) -> list[str]:
         if self.coefficient != 1:
@@ -216,9 +262,10 @@ class Z(StandardObservable):
         self, serialization_properties: OpenQASMSerializationProperties, target: QubitSet = None
     ) -> str:
         coef_prefix = f"{self.coefficient} * " if self.coefficient != 1 else ""
-        if target:
-            qubit_target = serialization_properties.format_target(int(target[0]))
-            return f"{coef_prefix}z({qubit_target})"
+        targets = target or self._targets
+        qubit_target = int(targets[0]) if targets else None
+        if qubit_target is not None:
+            return f"{coef_prefix}z({serialization_properties.format_target(qubit_target)})"
         return f"{coef_prefix}z all"
 
     def to_matrix(self) -> np.ndarray:
@@ -242,13 +289,13 @@ class TensorProduct(Observable):
             observables (list[Observable]): List of observables for tensor product
 
         Examples:
-            >>> t1 = Observable.Y() @ Observable.X()
+            >>> t1 = Observable.Y(0) @ Observable.X(1)
             >>> t1.to_matrix()
             array([[0.+0.j, 0.+0.j, 0.-0.j, 0.-1.j],
             [0.+0.j, 0.+0.j, 0.-1.j, 0.-0.j],
             [0.+0.j, 0.+1.j, 0.+0.j, 0.+0.j],
             [0.+1.j, 0.+0.j, 0.+0.j, 0.+0.j]])
-            >>> t2 = Observable.Z() @ t1
+            >>> t2 = Observable.Z(3) @ t1
             >>> t2.factors
             (Z('qubit_count': 1), Y('qubit_count': 1), X('qubit_count': 1))
 
@@ -276,7 +323,22 @@ class TensorProduct(Observable):
             f"{coefficient if coefficient != 1 else ''}"
             f"{'@'.join([obs.ascii_symbols[0] for obs in unscaled_factors])}"
         )
-        super().__init__(qubit_count=qubit_count, ascii_symbols=[display_name] * qubit_count)
+        all_targets = [factor.targets for factor in unscaled_factors]
+        if all(targets is None for targets in all_targets):
+            merged_targets = None
+        elif all(targets is not None for targets in all_targets):
+            flat_targets = [qubit for target in all_targets for qubit in target]
+            merged_targets = QubitSet(flat_targets)
+            if len(merged_targets) != len(flat_targets):
+                raise ValueError("Cannot have repeated target qubits")
+        else:
+            raise ValueError("Cannot mix factors with and without targets")
+
+        super().__init__(
+            qubit_count=qubit_count,
+            ascii_symbols=[display_name] * qubit_count,
+            targets=merged_targets,
+        )
         self._coef = coefficient
         self._factors = unscaled_factors
         self._factor_dimensions = tuple(
@@ -311,7 +373,7 @@ class TensorProduct(Observable):
     ) -> str:
         coef_prefix = f"{self.coefficient} * " if self.coefficient != 1 else ""
         factors = []
-        use_qubits = iter(target)
+        use_qubits = iter(target or self._targets)
         for obs in self._factors:
             obs_target = QubitSet()
             num_qubits = int(np.log2(obs.to_matrix().shape[0]))
@@ -433,7 +495,7 @@ class Sum(Observable):
                 observable for circuit diagrams. Defaults to `Hamiltonian`.
 
         Examples:
-            >>> t1 = -3 * Observable.Y() + 2 * Observable.X()
+            >>> t1 = -3 * Observable.Y(0) + 2 * Observable.X(0)
             Sum(X('qubit_count': 1), Y('qubit_count': 1))
             >>> t1.summands
             (X('qubit_count': 1), Y('qubit_count': 1))
@@ -447,7 +509,15 @@ class Sum(Observable):
 
         self._summands = tuple(flattened_observables)
         qubit_count = max(flattened_observables, key=lambda obs: obs.qubit_count).qubit_count
+        all_targets = [observable.targets for observable in flattened_observables]
+        if all(targets is None for targets in all_targets):
+            targets = None
+        elif all(targets is not None for targets in all_targets):
+            targets = all_targets
+        else:
+            raise ValueError("Cannot mix terms with and without targets")
         super().__init__(qubit_count=qubit_count, ascii_symbols=[display_name] * qubit_count)
+        self._targets = targets
 
     def __mul__(self, other: numbers.Number) -> Observable:
         """Scalar multiplication"""
@@ -464,8 +534,9 @@ class Sum(Observable):
     def _to_openqasm(
         self,
         serialization_properties: OpenQASMSerializationProperties,
-        target: list[QubitSet] | None = None,
+        target: list[QubitSetInput] | None = None,
     ) -> str:
+        target = target or self._targets
         if len(self.summands) != len(target):
             raise ValueError(
                 f"Invalid target of length {len(target)} for Sum with {len(self.summands)} terms"
@@ -524,21 +595,30 @@ class Hermitian(Observable):
     # Cache of eigenpairs
     _eigenpairs: ClassVar = {}
 
-    def __init__(self, matrix: np.ndarray, display_name: str = "Hermitian"):
+    def __init__(
+        self,
+        matrix: np.ndarray,
+        display_name: str = "Hermitian",
+        targets: QubitSetInput | None = None,
+    ):
         """Inits a `Hermitian`.
 
         Args:
             matrix (np.ndarray): Hermitian matrix that defines the observable.
             display_name (str): Name to use for an instance of this Hermitian matrix
                 observable for circuit diagrams. Defaults to `Hermitian`.
+            targets (QubitSetInput | None): The target qubits to measure the observable on.
+                If not provided, this needs to be provided from the enclosing result type.
+                Default: `None`.
 
         Raises:
             ValueError: If `matrix` is not a two-dimensional square matrix,
-                or has a dimension length that is not a positive power of 2,
-                or is not Hermitian.
+                is not Hermitian, or has a dimension length that is either not a positive power of 2
+                or, if targets is supplied, doesn't match the size of targets.
 
         Examples:
-            >>> Observable.Hermitian(matrix=np.array([[0, 1], [1, 0]]))
+            >>> observables.Hermitian(matrix=np.array([[0, 1],[1, 0]]), targets=[0])
+            >>> observables.Hermitian(matrix=np.array([[0, 1],[1, 0]]))
         """
         verify_quantum_operator_matrix_dimensions(matrix)
         self._matrix = np.array(matrix, dtype=complex)
@@ -552,10 +632,14 @@ class Hermitian(Observable):
             Gate.Unitary(matrix=eigendecomposition["eigenvectors"].conj().T),
         )
 
-        super().__init__(qubit_count=qubit_count, ascii_symbols=[display_name] * qubit_count)
+        super().__init__(
+            qubit_count=qubit_count, ascii_symbols=[display_name] * qubit_count, targets=targets
+        )
 
     def _unscaled(self) -> Observable:
-        return Hermitian(matrix=self._matrix, display_name=self.ascii_symbols[0])
+        return Hermitian(
+            matrix=self._matrix, display_name=self.ascii_symbols[0], targets=self._targets
+        )
 
     def _to_jaqcd(self) -> list[list[list[list[float]]]]:
         if self.coefficient != 1:
@@ -568,6 +652,7 @@ class Hermitian(Observable):
         self, serialization_properties: OpenQASMSerializationProperties, target: QubitSet = None
     ) -> str:
         coef_prefix = f"{self.coefficient} * " if self.coefficient != 1 else ""
+        target = target or self._targets
         if target:
             qubit_target = ", ".join([
                 serialization_properties.format_target(int(t)) for t in target