+++ src/braket/circuits/result_types.py
@@ -20,7 +20,6 @@ import braket.ir.jaqcd as ir
 from braket.circuits import circuit
 from braket.circuits.free_parameter import FreeParameter
 from braket.circuits.observable import Observable
-from braket.circuits.observables import Sum
 from braket.circuits.result_type import (
     ObservableParameterResultType,
     ObservableResultType,
@@ -95,7 +94,7 @@ class DensityMatrix(ResultType):
                 full density matrix is returned.
 
         Examples:
-            >>> ResultType.DensityMatrix(target=[0, 1])
+            >>> result_types.DensityMatrix(target=[0, 1])
         """
         self._target = QubitSet(target)
         ascii_symbols = ["DensityMatrix"] * len(self._target) if self._target else ["DensityMatrix"]
@@ -196,22 +195,19 @@ class AdjointGradient(ObservableParameterResultType):
 
 
         Examples:
-            >>> ResultType.AdjointGradient(observable=Observable.Z(),
+            >>> result_types.AdjointGradient(observable=observables.Z(0),
+                                        parameters=["alpha", "beta"])
+            >>> result_types.AdjointGradient(observable=observables.Z(),
                                         target=0, parameters=["alpha", "beta"])
 
-            >>> tensor_product = Observable.Y() @ Observable.Z()
-            >>> hamiltonian = Observable.Y() @ Observable.Z() + Observable.H()
-            >>> ResultType.AdjointGradient(
+            >>> tensor_product = observables.Y(0) @ observables.Z(1)
+            >>> hamiltonian = observables.Y(0) @ observables.Z(1) + observables.H(0)
+            >>> result_types.AdjointGradient(
             >>>     observable=tensor_product,
-            >>>     target=[[0, 1], [2]],
             >>>     parameters=["alpha", "beta"],
             >>> )
         """
-        if isinstance(observable, Sum):
-            target_qubits = reduce(QubitSet.union, map(QubitSet, target), QubitSet())
-        else:
-            target_qubits = QubitSet(target)
-
+        target_qubits = QubitSet(target if target is not None else observable.targets)
         super().__init__(
             ascii_symbols=[f"AdjointGradient({observable.ascii_symbols[0]})"] * len(target_qubits),
             observable=observable,
@@ -259,7 +255,7 @@ class AdjointGradient(ObservableParameterResultType):
         Examples:
             >>> alpha, beta = FreeParameter("alpha"), FreeParameter("beta")
             >>> circ = Circuit().h(0).h(1).rx(0, alpha).yy(0, 1, beta).adjoint_gradient(
-            >>>     observable=Observable.Z(), target=[0], parameters=[alpha, beta]
+            >>>     observable=observables.Z(0), parameters=[alpha, beta]
             >>> )
         """
         return ResultType.AdjointGradient(
@@ -286,7 +282,7 @@ class Amplitude(ResultType):
                 state is not a list of strings of '0' and '1'
 
         Examples:
-            >>> ResultType.Amplitude(state=["01", "10"])
+            >>> result_types.Amplitude(state=['01', '10'])
         """
         if (
             not state
@@ -365,7 +361,7 @@ class Probability(ResultType):
                 circuit.
 
         Examples:
-            >>> ResultType.Probability(target=[0, 1])
+            >>> result_types.Probability(target=[0, 1])
         """
         self._target = QubitSet(target)
         ascii_symbols = ["Probability"] * len(self._target) if self._target else ["Probability"]
@@ -450,16 +446,18 @@ class Expectation(ObservableResultType):
 
         Args:
             observable (Observable): the observable for the result type
-            target (QubitSetInput | None): Target qubits that the
-                result type is requested for. Default is `None`, which means the observable must
-                operate only on 1 qubit and it is applied to all qubits in parallel.
-
+            target (QubitSetInput | None): Target qubits that the result type is requested for.
+                If not provided, the observable's target will be used instead. If neither exist,
+                then it is applied to all qubits in parallel; in this case the observable must
+                operate only on 1 qubit.
+                Default: `None`.
 
         Examples:
-            >>> ResultType.Expectation(observable=Observable.Z(), target=0)
+            >>> result_types.Expectation(observable=observables.Z(0))
+            >>> result_types.Expectation(observable=observables.Z(), target=0)
 
-            >>> tensor_product = Observable.Y() @ Observable.Z()
-            >>> ResultType.Expectation(observable=tensor_product, target=[0, 1])
+            >>> tensor_product = observables.Y(0) @ observables.Z(1)
+            >>> result_types.Expectation(observable=tensor_product)
         """
         super().__init__(
             ascii_symbols=[f"Expectation({obs_ascii})" for obs_ascii in observable.ascii_symbols],
@@ -497,7 +495,7 @@ class Expectation(ObservableResultType):
             ResultType: expectation as a requested result type
 
         Examples:
-            >>> circ = Circuit().expectation(observable=Observable.Z(), target=0)
+            >>> circ = Circuit().expectation(observable=observables.Z(0))
         """
         return ResultType.Expectation(observable=observable, target=target)
 
@@ -522,15 +520,18 @@ class Sample(ObservableResultType):
 
         Args:
             observable (Observable): the observable for the result type
-            target (QubitSetInput | None): Target qubits that the
-                result type is requested for. Default is `None`, which means the observable must
-                operate only on 1 qubit and it is applied to all qubits in parallel.
+            target (QubitSetInput | None): Target qubits that the result type is requested for.
+                If not provided, the observable's target will be used instead. If neither exist,
+                then it is applied to all qubits in parallel; in this case the observable must
+                operate only on 1 qubit.
+                Default: `None`.
 
         Examples:
-            >>> ResultType.Sample(observable=Observable.Z(), target=0)
+            >>> result_types.Sample(observable=observables.Z(0))
+            >>> result_types.Sample(observable=observables.Z(), target=0)
 
-            >>> tensor_product = Observable.Y() @ Observable.Z()
-            >>> ResultType.Sample(observable=tensor_product, target=[0, 1])
+            >>> tensor_product = observables.Y(0) @ observables.Z(1)
+            >>> result_types.Sample(observable=tensor_product)
         """
         super().__init__(
             ascii_symbols=[f"Sample({obs_ascii})" for obs_ascii in observable.ascii_symbols],
@@ -568,7 +569,7 @@ class Sample(ObservableResultType):
             ResultType: sample as a requested result type
 
         Examples:
-            >>> circ = Circuit().sample(observable=Observable.Z(), target=0)
+            >>> circ = Circuit().sample(observable=observables.Z(0))
         """
         return ResultType.Sample(observable=observable, target=target)
 
@@ -594,19 +595,22 @@ class Variance(ObservableResultType):
 
         Args:
             observable (Observable): the observable for the result type
-            target (QubitSetInput | None): Target qubits that the
-                result type is requested for. Default is `None`, which means the observable must
-                operate only on 1 qubit and it is applied to all qubits in parallel.
+            target (QubitSetInput | None): Target qubits that the result type is requested for.
+                If not provided, the observable's target will be used instead. If neither exist,
+                then it is applied to all qubits in parallel; in this case the observable must
+                operate only on 1 qubit.
+                Default: `None`.
 
         Raises:
             ValueError: If the observable's qubit count does not equal the number of target
                 qubits, or if `target=None` and the observable's qubit count is not 1.
 
         Examples:
-            >>> ResultType.Variance(observable=Observable.Z(), target=0)
+            >>> result_types.Variance(observable=observables.Z(0))
+            >>> result_types.Variance(observable=observables.Z(), target=0)
 
-            >>> tensor_product = Observable.Y() @ Observable.Z()
-            >>> ResultType.Variance(observable=tensor_product, target=[0, 1])
+            >>> tensor_product = observables.Y(0) @ observables.Z(1)
+            >>> result_types.Variance(observable=tensor_product)
         """
         super().__init__(
             ascii_symbols=[f"Variance({obs_ascii})" for obs_ascii in observable.ascii_symbols],
@@ -644,7 +648,7 @@ class Variance(ObservableResultType):
             ResultType: variance as a requested result type
 
         Examples:
-            >>> circ = Circuit().variance(observable=Observable.Z(), target=0)
+            >>> circ = Circuit().variance(observable=observables.Z(0))
         """
         return ResultType.Variance(observable=observable, target=target)