+++ src/braket/pulse/waveforms.py
@@ -504,6 +504,7 @@ class ErfSquareWaveform(Waveform, Parameterizable):
         length: float | FreeParameterExpression,
         width: float | FreeParameterExpression,
         sigma: float | FreeParameterExpression,
+        off_center: float | FreeParameterExpression = 0,
         amplitude: float | FreeParameterExpression = 1,
         zero_at_edges: bool = False,
         id: str | None = None,
@@ -517,22 +518,26 @@ class ErfSquareWaveform(Waveform, Parameterizable):
         height. The waveform is scaled such that its maximum is equal to `amplitude`.
 
         Args:
-            length (Union[float, FreeParameterExpression]): Duration (in seconds) from the start
+            length (float | FreeParameterExpression): Duration (in seconds) from the start
                 to the end of the waveform.
-            width (Union[float, FreeParameterExpression]): Duration (in seconds) between the
+            width (float | FreeParameterExpression): Duration (in seconds) between the
                 half height of the two edges.
-            sigma (Union[float, FreeParameterExpression]): A characteristic time of how quickly
+            sigma (float | FreeParameterExpression): A characteristic time of how quickly
                 the edges rise and fall.
-            amplitude (Union[float, FreeParameterExpression]): The amplitude of the waveform
+            off_center (float | FreeParameterExpression): Shift the smoothed square waveform
+                earlier or later in time. When positive, the smoothed square is shifted later
+                (to the right), otherwise earlier (to the left). Defaults to 0.
+            amplitude (float | FreeParameterExpression): The amplitude of the waveform
                 envelope. Defaults to 1.
             zero_at_edges (bool): Whether the waveform is scaled such that it has zero value at the
                 edges. Defaults to False.
-            id (Optional[str]): The identifier used for declaring this waveform. A random string of
+            id (str | None): The identifier used for declaring this waveform. A random string of
                 ascii characters is assigned by default.
         """
         self.length = length
         self.width = width
         self.sigma = sigma
+        self.off_center = off_center
         self.amplitude = amplitude
         self.zero_at_edges = zero_at_edges
         self.id = id or _make_identifier_name()
@@ -540,8 +545,8 @@ class ErfSquareWaveform(Waveform, Parameterizable):
     def __repr__(self) -> str:
         return (
             f"ErfSquareWaveform('id': {self.id}, 'length': {self.length}, "
-            f"'width': {self.width}, 'sigma': {self.sigma}, 'amplitude': {self.amplitude}, "
-            f"'zero_at_edges': {self.zero_at_edges})"
+            f"'width': {self.width}, 'sigma': {self.sigma}, 'off_center': {self.off_center}, "
+            f"'amplitude': {self.amplitude}, 'zero_at_edges': {self.zero_at_edges})"
         )
 
     @property
@@ -549,14 +554,14 @@ class ErfSquareWaveform(Waveform, Parameterizable):
         """Returns the parameters associated with the object, either unbound free parameter
         expressions or bound values.
         """
-        return [self.length, self.width, self.sigma, self.amplitude]
+        return [self.length, self.width, self.sigma, self.off_center, self.amplitude]
 
     def bind_values(self, **kwargs: FreeParameter | str) -> ErfSquareWaveform:
         """Takes in parameters and returns an object with specified parameters
         replaced with their values.
 
         Args:
-            **kwargs (Union[FreeParameter, str]): Arbitrary keyword arguments.
+            **kwargs (FreeParameter | str): Arbitrary keyword arguments.
 
         Returns:
             ErfSquareWaveform: A copy of this waveform with the requested parameters bound.
@@ -565,6 +570,7 @@ class ErfSquareWaveform(Waveform, Parameterizable):
             "length": subs_if_free_parameter(self.length, **kwargs),
             "width": subs_if_free_parameter(self.width, **kwargs),
             "sigma": subs_if_free_parameter(self.sigma, **kwargs),
+            "off_center": subs_if_free_parameter(self.off_center, **kwargs),
             "amplitude": subs_if_free_parameter(self.amplitude, **kwargs),
             "zero_at_edges": self.zero_at_edges,
             "id": self.id,
@@ -576,6 +582,7 @@ class ErfSquareWaveform(Waveform, Parameterizable):
             self.length,
             self.width,
             self.sigma,
+            self.off_center,
             self.amplitude,
             self.zero_at_edges,
             self.id,
@@ -583,6 +590,7 @@ class ErfSquareWaveform(Waveform, Parameterizable):
             other.length,
             other.width,
             other.sigma,
+            other.off_center,
             other.amplitude,
             other.zero_at_edges,
             other.id,
@@ -600,6 +608,7 @@ class ErfSquareWaveform(Waveform, Parameterizable):
                 ("length", duration),
                 ("width", duration),
                 ("sigma", duration),
+                ("off_center", duration),
                 ("amplitude", float64),
                 ("zero_at_edges", bool_),
             ],
@@ -609,6 +618,7 @@ class ErfSquareWaveform(Waveform, Parameterizable):
                 self.length,
                 self.width,
                 self.sigma,
+                self.off_center,
                 self.amplitude,
                 self.zero_at_edges,
             ),
@@ -625,8 +635,8 @@ class ErfSquareWaveform(Waveform, Parameterizable):
             np.ndarray: The sample amplitudes for this waveform.
         """
         sample_range = np.arange(0, self.length, dt)
-        t1 = (self.length - self.width) / 2
-        t2 = (self.length + self.width) / 2
+        t1 = (self.length - self.width) / 2 + self.off_center
+        t2 = (self.length + self.width) / 2 + self.off_center
         samples = (
             sp.special.erf((sample_range - t1) / self.sigma)
             + sp.special.erf(-(sample_range - t2) / self.sigma)