+++ test/unit_tests/braket/circuits/test_observables.py
@@ -61,100 +61,116 @@ def test_to_ir(testobject, gateobject, expected_ir, basis_rotation_gates, eigenv
 
 
 @pytest.mark.parametrize(
-    "observable, serialization_properties, target, expected_ir",
+    "observable, observable_with_targets, serialization_properties, target, expected_ir",
     [
         (
             Observable.I(),
+            Observable.I(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [3],
             "i(q[3])",
         ),
         (
             Observable.I(),
+            Observable.I(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3],
             "i($3)",
         ),
         (
             Observable.I(),
+            None,
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             None,
             "i all",
         ),
         (
             Observable.X(),
+            Observable.X(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [3],
             "x(q[3])",
         ),
         (
             Observable.X(),
+            Observable.X(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3],
             "x($3)",
         ),
         (
             Observable.X(),
+            None,
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             None,
             "x all",
         ),
         (
             Observable.Y(),
+            Observable.Y(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [3],
             "y(q[3])",
         ),
         (
             Observable.Y(),
+            Observable.Y(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3],
             "y($3)",
         ),
         (
             Observable.Y(),
+            None,
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             None,
             "y all",
         ),
         (
             Observable.Z(),
+            Observable.Z(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [3],
             "z(q[3])",
         ),
         (
             Observable.Z(),
+            Observable.Z(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3],
             "z($3)",
         ),
         (
             Observable.Z(),
+            None,
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             None,
             "z all",
         ),
         (
             Observable.H(),
+            Observable.H(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [3],
             "h(q[3])",
         ),
         (
             Observable.H(),
+            Observable.H(3),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3],
             "h($3)",
         ),
         (
             Observable.H(),
+            None,
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             None,
             "h all",
         ),
         (
             Observable.Hermitian(np.eye(4)),
+            Observable.Hermitian(np.eye(4), targets=[1, 2]),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [1, 2],
             "hermitian([[1+0im, 0im, 0im, 0im], [0im, 1+0im, 0im, 0im], "
@@ -162,6 +178,7 @@ def test_to_ir(testobject, gateobject, expected_ir, basis_rotation_gates, eigenv
         ),
         (
             Observable.Hermitian(np.eye(4)),
+            Observable.Hermitian(np.eye(4), targets=[1, 2]),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [1, 2],
             "hermitian([[1+0im, 0im, 0im, 0im], [0im, 1+0im, 0im, 0im], "
@@ -169,36 +186,42 @@ def test_to_ir(testobject, gateobject, expected_ir, basis_rotation_gates, eigenv
         ),
         (
             Observable.Hermitian(np.eye(2)),
+            None,
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             None,
             "hermitian([[1+0im, 0im], [0im, 1+0im]]) all",
         ),
         (
             Observable.H() @ Observable.Z(),
+            Observable.H(3) @ Observable.Z(0),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [3, 0],
             "h(q[3]) @ z(q[0])",
         ),
         (
             Observable.H() @ Observable.Z(),
+            Observable.H(3) @ Observable.Z(0),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3, 0],
             "h($3) @ z($0)",
         ),
         (
             Observable.H() @ Observable.Z() @ Observable.I(),
+            Observable.H(3) @ Observable.Z(0) @ Observable.I(1),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [3, 0, 1],
             "h(q[3]) @ z(q[0]) @ i(q[1])",
         ),
         (
             Observable.H() @ Observable.Z() @ Observable.I(),
+            Observable.H(3) @ Observable.Z(0) @ Observable.I(1),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3, 0, 1],
             "h($3) @ z($0) @ i($1)",
         ),
         (
             Observable.Hermitian(np.eye(4)) @ Observable.I(),
+            Observable.Hermitian(np.eye(4), targets=[3, 0]) @ Observable.I(1),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.VIRTUAL),
             [3, 0, 1],
             "hermitian([[1+0im, 0im, 0im, 0im], [0im, 1+0im, 0im, 0im], "
@@ -207,6 +230,7 @@ def test_to_ir(testobject, gateobject, expected_ir, basis_rotation_gates, eigenv
         ),
         (
             Observable.I() @ Observable.Hermitian(np.eye(4)),
+            Observable.I(3) @ Observable.Hermitian(np.eye(4), targets=[0, 1]),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3, 0, 1],
             "i($3) @ "
@@ -214,25 +238,15 @@ def test_to_ir(testobject, gateobject, expected_ir, basis_rotation_gates, eigenv
             "[0im, 0im, 1+0im, 0im], [0im, 0im, 0im, 1+0im]]) $0, $1",
         ),
         (
-            (2 * Observable.Z()) @ (3 * Observable.H()),
-            OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
-            [3, 3],
-            "6 * z($3) @ h($3)",
-        ),
-        (
-            (2 * Observable.Z()) @ (3 * Observable.H()) @ (2 * Observable.Y()),
-            OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
-            [3, 3, 1],
-            "12 * z($3) @ h($3) @ y($1)",
-        ),
-        (
             3 * (2 * Observable.Z()),
+            3 * (2 * Observable.Z(3)),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3],
             "6 * z($3)",
         ),
         (
             (2 * Observable.I()) @ (2 * Observable.Hermitian(np.eye(4))),
+            (2 * Observable.I(3)) @ (2 * Observable.Hermitian(np.eye(4), targets=[0, 1])),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [3, 0, 1],
             "4 * i($3) @ "
@@ -241,55 +255,84 @@ def test_to_ir(testobject, gateobject, expected_ir, basis_rotation_gates, eigenv
         ),
         (
             Observable.Z() + 2 * Observable.H(),
+            Observable.Z(3) + 2 * Observable.H(4),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [[3], [4]],
             "z($3) + 2 * h($4)",
         ),
         (
             3 * (Observable.H() + 2 * Observable.X()),
+            3 * (Observable.H(3) + 2 * Observable.X(0)),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [[3], [0]],
             "3 * h($3) + 6 * x($0)",
         ),
         (
             3 * (Observable.H() + 2 * Observable.H()),
+            3 * (Observable.H(3) + 2 * Observable.H(3)),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [[3], [3]],
             "3 * h($3) + 6 * h($3)",
         ),
         (
             3 * (Observable.H() + 2 * Observable.H()),
+            3 * (Observable.H(3) + 2 * Observable.H(5)),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [[3], [5]],
             "3 * h($3) + 6 * h($5)",
         ),
         (
             (2 * Observable.Y()) @ (3 * Observable.I()) + 0.75 * Observable.Y() @ Observable.Z(),
+            (2 * Observable.Y(0)) @ (3 * Observable.I(1))
+            + 0.75 * Observable.Y(0) @ Observable.Z(1),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [[0, 1], [0, 1]],
             "6 * y($0) @ i($1) + 0.75 * y($0) @ z($1)",
         ),
         (
             (-2 * Observable.Y()) @ (3 * Observable.I()) + -0.75 * Observable.Y() @ Observable.Z(),
+            (-2 * Observable.Y(0)) @ (3 * Observable.I(1))
+            + -0.75 * Observable.Y(0) @ Observable.Z(1),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [[0, 1], [0, 1]],
             "-6 * y($0) @ i($1) - 0.75 * y($0) @ z($1)",
         ),
         (
             4 * (2 * Observable.Z() + 2 * (3 * Observable.X() @ (2 * Observable.Y()))),
+            4 * (2 * Observable.Z(0) + 2 * (3 * Observable.X(1) @ (2 * Observable.Y(2)))),
             OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
             [[0], [1, 2]],
             "8 * z($0) + 48 * x($1) @ y($2)",
         ),
+        (
+            4 * (2 * Observable.Z(0) + 2 * (3 * Observable.X(1) @ (2 * Observable.Y(2)))),
+            None,
+            OpenQASMSerializationProperties(qubit_reference_type=QubitReferenceType.PHYSICAL),
+            [[5], [4, 3]],
+            "8 * z($5) + 48 * x($4) @ y($3)",
+        ),
     ],
 )
-def test_observables_to_ir_openqasm(observable, serialization_properties, target, expected_ir):
+def test_observables_to_ir_openqasm(
+    observable,
+    observable_with_targets,
+    serialization_properties,
+    target,
+    expected_ir,
+):
     assert (
         observable.to_ir(
             target, ir_type=IRType.OPENQASM, serialization_properties=serialization_properties
         )
         == expected_ir
     )
+    if observable_with_targets:
+        assert (
+            observable_with_targets.to_ir(
+                None, ir_type=IRType.OPENQASM, serialization_properties=serialization_properties
+            )
+            == expected_ir
+        )
 
 
 @pytest.mark.parametrize(
@@ -459,6 +502,11 @@ def test_hermitian_eigenvalues(matrix, eigenvalues):
     compare_eigenvalues(Observable.Hermitian(matrix=matrix), eigenvalues)
 
 
+def test_hermitian_matrix_target_mismatch():
+    with pytest.raises(ValueError):
+        Observable.Hermitian(np.eye(4), targets=[0, 1, 2])
+
+
 def test_flattened_tensor_product():
     observable_one = Observable.Z() @ Observable.Y()
     observable_two = Observable.X() @ Observable.H()
@@ -616,6 +664,16 @@ def test_tensor_product_basis_rotation_gates(observable, basis_rotation_gates):
     assert observable.basis_rotation_gates == basis_rotation_gates
 
 
+def test_tensor_product_repeated_qubits():
+    with pytest.raises(ValueError):
+        (2 * Observable.Z(3)) @ (3 * Observable.H(3))
+
+
+def test_tensor_product_with_and_without_targets():
+    with pytest.raises(ValueError):
+        (2 * Observable.Z(3)) @ (3 * Observable.H())
+
+
 def test_observable_from_ir_tensor_product():
     expected_observable = Observable.TensorProduct([Observable.Z(), Observable.I(), Observable.X()])
     actual_observable = observable_from_ir(["z", "i", "x"])
@@ -718,3 +776,8 @@ def test_unscaled_tensor_product():
     observable = 3 * ((2 * Observable.X()) @ (5 * Observable.Y()))
     assert observable == 30 * (Observable.X() @ Observable.Y())
     assert observable._unscaled() == Observable.X() @ Observable.Y()
+
+
+def test_sum_with_and_without_targets():
+    with pytest.raises(ValueError):
+        Observable.X() + 3 * Observable.Y(4)